# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification Service API (Public)",
        "description": "FCM & APNs Token Management - AUTH REMOVED",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/notifications/tokens": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Register token (Publicly)",
                "operationId": "1577941aa2b73c88ac64702f161d8e49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "token",
                                    "provider"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "fcm_token_123"
                                    },
                                    "provider": {
                                        "type": "string",
                                        "enum": [
                                            "fcm",
                                            "apns"
                                        ]
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "example": "Web/Android"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/notifications/send": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send Push (Publicly)",
                "operationId": "dd48e861fa183e13341ca747324f1350",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "body"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sent"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Notifications",
            "description": "Notifications"
        }
    ]
}